backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pytest==7.4.3
pytest-django==4.7.0
gunicorn==21.2.0
python-magic==0.4.27 